+++ ServicesRegister/src/main/java/org/example/controllers/MaintenanceController.java
@@ -5,9 +5,10 @@ import org.example.documents.Maintenance;
 import org.example.safety.Constants;
 import org.example.services.MaintenanceServices;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
+import org.springframework.http.*;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.client.RestTemplate;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -17,29 +18,105 @@ public class MaintenanceController implements CRUDController<Maintenance, Mainte
     @Autowired
     private MaintenanceServices services;
 
+    private RestTemplate restTemplate = new RestTemplate();
+    private HttpHeaders headers = new HttpHeaders();
+
     @Override
     @PostMapping(Constants.MAINTENANCE)
     public ResponseEntity<Maintenance> create(@RequestHeader(value = "Authorization") String token, @RequestBody MaintenanceDTO maintenanceDTO) {
-        Maintenance m = services.save(maintenanceDTO);
-        return ResponseEntity.status(HttpStatus.CREATED).body(m);
+        try{
+            headers.set("Content-Type", "application/json");
+
+            HttpEntity<Boolean> entity = new HttpEntity<>(headers);
+
+            ResponseEntity<Boolean> response = restTemplate.exchange(
+                    Constants.VALIDATE,
+                    HttpMethod.GET,
+                    entity,
+                    Boolean.class
+            );
+
+            if (response.getStatusCode().is2xxSuccessful() && response.getBody() != null) {
+                Maintenance m = services.save(maintenanceDTO);
+
+                return ResponseEntity.status(HttpStatus.CREATED).body(m);
+            }
+        }catch (Exception e){
+            System.err.println("Error trying to validate User\n" + e.getMessage());
+        }
+        throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Incorrect User or Password");
     }
 
     @Override
     @GetMapping(Constants.MAINTENANCE)
     public ResponseEntity<List<Maintenance>> findAll(@RequestHeader(value = "Authorization") String token) {
-        return ResponseEntity.ok(services.findAll());
+        try{
+            headers.set("Content-Type", "application/json");
+
+            HttpEntity<Boolean> entity = new HttpEntity<>(headers);
+
+            ResponseEntity<Boolean> response = restTemplate.exchange(
+                    Constants.VALIDATE,
+                    HttpMethod.GET,
+                    entity,
+                    Boolean.class
+            );
+
+            if (response.getStatusCode().is2xxSuccessful() && response.getBody() != null) {
+                return ResponseEntity.ok(services.findAll());
+            }
+        }catch (Exception e){
+            System.err.println("Error trying to validate User\n" + e.getMessage());
+        }
+        throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Incorrect User or Password");
     }
 
     @Override
     @GetMapping(Constants.MAINTENANCE + "/{id}")
     public ResponseEntity<Optional<Maintenance>> findById(@RequestHeader(value = "Authorization") String token, @PathVariable String id) {
-        return ResponseEntity.ok(services.findByID(id));
+        try{
+            headers.set("Content-Type", "application/json");
+
+            HttpEntity<Boolean> entity = new HttpEntity<>(headers);
+
+            ResponseEntity<Boolean> response = restTemplate.exchange(
+                    Constants.VALIDATE,
+                    HttpMethod.GET,
+                    entity,
+                    Boolean.class
+            );
+
+            if (response.getStatusCode().is2xxSuccessful() && response.getBody() != null) {
+                return ResponseEntity.ok(services.findByID(id));
+            }
+        }catch (Exception e){
+            System.err.println("Error trying to validate User\n" + e.getMessage());
+        }
+        throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Incorrect User or Password");
     }
 
     @Override
     @DeleteMapping(Constants.MAINTENANCE + "/{id}")
     public ResponseEntity<Void> deleteById(@RequestHeader(value = "Authorization") String token, @PathVariable String id) {
-        services.deleteByID(id);
-        return ResponseEntity.noContent().build();
+        try{
+            headers.set("Content-Type", "application/json");
+
+            HttpEntity<Boolean> entity = new HttpEntity<>(headers);
+
+            ResponseEntity<Boolean> response = restTemplate.exchange(
+                    Constants.VALIDATE,
+                    HttpMethod.GET,
+                    entity,
+                    Boolean.class
+            );
+
+            if (response.getStatusCode().is2xxSuccessful() && response.getBody() != null) {
+                services.deleteByID(id);
+                return ResponseEntity.noContent().build();
+            }
+        }catch (Exception e){
+            System.err.println("Error trying to validate User\n" + e.getMessage());
+        }
+        throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Incorrect User or Password");
     }
 }
